14ПР

+++ controllers/cards.js
@@ -1,43 +1,51 @@
 const Card = require('../models/card');
-const errorStatus = require('../utils/errorStatus');
+
+const NotFound = require('../errors/not-found-err');
+const BadRequest = require('../errors/bad-request-err');
+const ForbiddenError = require('../errors/forbidden-err');
 
 // Получить все карточки.
 const getCards = (req, res) => {
   Card.find({})
     .then((cards) => res.send(cards))
-    .catch(() => res.status(errorStatus.SERVER_ERROR).send({ message: 'На сервере произошла ошибка' }));
+    .catch(next);
 };
 
 // Создать карточку.
 const createCard = (req, res) => {
   const { name, link } = req.body;
-  const { _id } = req.user;
-  Card.create({ name, link, owner: _id })
+  const owner = req.user._id;
+  Card.create({ name, link, owner })
     .then((card) => res.send({ data: card }))
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        res.status(errorStatus.BAD_REQUEST).send({ message: 'Переданы некорректные данные' });
-      } else {
-        res.status(errorStatus.SERVER_ERROR).send({ message: 'На сервере произошла ошибка' });
+        return next(new BadRequest('Переданы некорректные данные'));
       }
+      return next(err);
     });
 };
 
 // Удалить карточку.
 const deleteCard = (req, res) => {
-  Card.findByIdAndDelete(req.params.cardId)
+  Card.findById(req.params.cardId)
     .then((card) => {
       if (!card) {
-        res.status(errorStatus.NOT_FOUND).send({ message: 'Карточка не найдена' });
-        return;
+        throw new NotFound('Карточка не найдена');
       }
-      res.send(card);
+      if (String(card.owner) !== req.user._id) {
+        throw new ForbiddenError('Удалять чужие карточки запрещено');
+      }
+
+      return Card.findByIdAndDelete(req.params.cardId)
+        .then(() => {
+          res.status(200).send({ message: 'Карточка успешно удалена' });
+        });
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(errorStatus.BAD_REQUEST).send({ message: 'Переданы некорректные данные' });
+        next(new BadRequest('Неверный запрос'));
       } else {
-        res.status(500).send({ message: 'На сервере произошла ошибка' });
+        next(err);
       }
     });
 };
@@ -47,16 +55,15 @@ const likeCard = (req, res) => {
   Card.findByIdAndUpdate(req.params.cardId, { $addToSet: { likes: req.user._id } }, { new: true })
     .then((card) => {
       if (!card) {
-        return res.status(errorStatus.NOT_FOUND).send({ message: 'Карточка не найдена' });
+        return next(new NotFound('Карточка не найдена'));
       }
       return res.send(card);
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(errorStatus.BAD_REQUEST).send({ message: 'Переданы некорректные данные' });
-      } else {
-        res.status(errorStatus.SERVER_ERROR).send({ message: 'На сервере произошла ошибка' });
+        return next(new BadRequest('Передан некорректный id карточки'));
       }
+      return next(err);
     });
 };
 
@@ -65,16 +72,15 @@ const deleteLike = (req, res) => {
   Card.findByIdAndUpdate(req.params.cardId, { $pull: { likes: req.user._id } }, { new: true })
     .then((card) => {
       if (!card) {
-        return res.status(errorStatus.NOT_FOUND).send({ message: 'Карточка не найдена' });
+        return next(new NotFound('Карточка не найдена'));
       }
       return res.send(card);
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(errorStatus.BAD_REQUEST).send({ message: 'Переданы некорректные данные' });
-      } else {
-        res.status(errorStatus.BAD_REQUEST).send({ message: 'На сервере произошла ошибка' });
+        return next(new BadRequest('Некорректный id карточки'));
       }
+      return next(err);
     });
 };
